requirements.txt
black==22.10.0
eth-brownie>=1.19.2,<2.0.0